move : 상수 파일 공통으로 이동
