+++ src/pages/MainPage.tsx
@@ -1,22 +1,25 @@
+import { useRecoilValue } from 'recoil';
 import { styled } from 'styled-components';
 
+import { APPLY_TYPE } from '../views/@common/utils/constants';
 import Banner from '../views/MainPage/components/Banner';
 import Contents from '../views/MainPage/components/Contents';
 import ReceivedOffer from '../views/MainPage/components/ReceivedOffer';
 import StatusBarForiOS from '../views/MainPage/components/StatusBarForiOS';
 import TopSheet from '../views/MainPage/components/TopSheet';
-import { APPLY_TYPE, USER_TYPE } from '../views/MainPage/constants/constants';
+
+import { userTypeState } from '@/recoil/atoms/signUpState';
 
 const MainPage = () => {
-  const userType = USER_TYPE.GUEST;
-  const applyType = APPLY_TYPE.RECEIVED;
+  const userType = useRecoilValue(userTypeState);
+  const applyType = APPLY_TYPE.NOT_YET;
 
   return (
     <MainPageLayout>
       <StatusBarForiOS />
-      <TopSheet userType={userType} applyType={applyType} />
+      <TopSheet applyType={applyType} />
       <Banner />
-      {userType === USER_TYPE.GUEST ? <Contents /> : <ReceivedOffer applyType={applyType} />}
+      {!userType ? <Contents /> : <ReceivedOffer applyType={applyType} />}
     </MainPageLayout>
   );
 };
