+++ src/views/@common/utils/constants.ts
@@ -1,9 +1,3 @@
-export const USER_TYPE = {
-  GUEST: 0,
-  DESIGNER: 1,
-  MODEL: 2,
-};
-
 export const APPLY_TYPE = {
   NOT_YET: 0,
   WAITING: 1,
