+++ src/views/@common/utils/userType.ts
@@ -1,0 +1,4 @@
+export const USER_TYPE = {
+  DESIGNER: 'designer',
+  MODEL: 'model',
+};
