+++ src/views/SignUpPage/components/VerifyPhoneNumber.tsx
@@ -4,10 +4,10 @@ import { styled } from 'styled-components';
 
 import Button from '../../@common/components/Button';
 import ProgressBar from '../../@common/components/ProgressBar';
+import { USER_TYPE } from '../../@common/utils/userType';
 import { HELPER_MESSAGE, PLACE_HOLDER_MESSAGE } from '../constants/message';
 import { STATUS } from '../constants/requestStatus';
 import { STEP, TOTAL_STEP } from '../constants/step';
-import { USER_TYPE } from '../constants/userType';
 import useInterval from '../hooks/useInterval';
 
 import Field from './Field';
