--- src/views/SignUpPage/constants/userType.ts
@@ -1,4 +1,0 @@
-export const USER_TYPE = {
-  DESIGNER: 0,
-  MODEL: 1,
-};
